# Code/Backend/src/main/resources/swagger/swagger.yaml
swagger: "2.0"
info:
  title: SimplePSUCalculator
  version: "0.1"
  description: SimplePSUCalculator - api

produces: ["application/json"]

tags:
  - name: Motherboard
  - name: Processor
  - name: Videocard

paths:

  "/processors":
    get:
      tags: [Processor]
      operationId: getProcessors
      security: []
      parameters:
        - "$ref": "#/parameters/CpuVendorParameter"
        - "$ref": "#/parameters/CpuFamilyParameter"
        - "$ref": "#/parameters/CpuSocketParameter"
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              "$ref": "#/definitions/CentralProcessingUnitModel"
        "400":
          description: Bad Request

  "/processors/filters":
    get:
      tags: [Processor]
      operationId: getProcessorFilters
      security: []
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/definitions/CentralProcessingUnitFiltersModel"
        "400":
          description: Bad Request

  "/motherboards":
    get:
      tags: [Motherboard]
      operationId: getMotherboards
      security: []
      parameters:
        - "$ref": "#/parameters/SupportedCpuSocketParameter"
        - "$ref": "#/parameters/MotherboardVendorParameter"
        - "$ref": "#/parameters/MotherboardChipsetParameter"
        - "$ref": "#/parameters/MotherboardFormParameter"
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              "$ref": "#/definitions/MotherboardModel"
        "400":
          description: Bad Request

  "/motherboard/filters":
    get:
      tags: [Motherboard]
      operationId: getMotherboardFilters
      security: []
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/definitions/MotherboardFiltersModel"
        "400":
          description: Bad Request

  "/videocards":
    get:
      tags: [Videocard]
      operationId: getVideocards
      security: []
      parameters:
        - "$ref": "#/parameters/VideocardGpuVendorParameter"
        - "$ref": "#/parameters/VideocardVendorParameter"
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              "$ref": "#/definitions/VideocardModel"
        "400":
          description: Bad Request

  "/videocards/filters":
    get:
      tags: [Videocard]
      operationId: getVideocardFilters
      security: []
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/definitions/VideocardFiltersModel"
        "400":
          description: Bad Request

definitions:
  MotherboardModel:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      vendor:
        type: string
      socket:
        type: string
      chipset:
        type: string
      form:
        type: string
      typeOfMemory:
        type: string
      numOfMemorySlots:
        type: integer
      maxMemoryVolume:
        type: integer
      numOfMemoryChannels:
        type: integer
      builtInGraphics:
        type: boolean
      yearOfIssue:
        type: string
      price:
        type: number
        format: double
      imageURL:
        type: string

  CentralProcessingUnitModel:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      vendor:
        type: string
      family:
        type: string
      codename:
        type: string
      socket:
        type: string
      numOfCores:
        type: integer
      numOfThreads:
        type: integer
      clockFrequency:
        type: number
        format: double
      maxClockFrequency:
        type: number
        format: double
      tdp:
        type: integer
      builtInGraphics:
        type: boolean
      yearOfIssue:
        type: string
      price:
        type: number
        format: double
      imageURL:
        type: string

  PowerSupplyUnitModel:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      vendor:
        type: string
      power:
        type: integer
      efficiency:
        type: integer
      certificate:
        type: boolean
      price:
        type: number
        format: double
      imageURL:
        type: string

  VideocardModel:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      vendor:
        type: string
      gpuVendor:
        type: string
      clockFrequency:
        type: number
        format: double
      maxClockFrequency:
        type: number
        format: double
      tdp:
        type: integer
      memoryVolume:
        type: number
        format: double
      typeOfMemory:
        type: string
      vgaSupport:
        type: boolean
      dviSupport:
        type: boolean
      hdmiSupport:
        type: boolean
      displayPortSupport:
        type: boolean
      yearOfIssue:
        type: string
      price:
        type: number
        format: double
      imageURL:
        type: string

  CentralProcessingUnitFiltersModel:
    type: object
    properties:
      vendors:
        type: array
        items:
          type: string
      families:
        type: array
        items:
          type: string
      sockets:
        type: array
        items:
          type: string

  MotherboardFiltersModel:
    type: object
    properties:
      vendors:
        type: array
        items:
          type: string
      chipsets:
        type: array
        items:
          type: string
      forms:
        type: array
        items:
          type: string

  VideocardFiltersModel:
    type: object
    properties:
      gpuVendors:
        type: array
        items:
          type: string
      vendors:
        type: array
        items:
          type: string

parameters:
  CpuVendorParameter:
    name: cpuVendor
    in: query
    type: string
    required: false

  CpuFamilyParameter:
    name: cpuFamily
    in: query
    type: string
    required: false

  CpuSocketParameter:
    name: cpuSocket
    in: query
    type: string
    required: false

  SupportedCpuSocketParameter:
    name: supportedCpuSocket
    in: query
    type: string
    required: true

  MotherboardVendorParameter:
    name: motherboardVendor
    in: query
    type: string
    required: false

  MotherboardChipsetParameter:
    name: motherboardChipset
    in: query
    type: string
    required: false

  MotherboardFormParameter:
    name: motherboardForm
    in: query
    type: string
    required: false

  VideocardGpuVendorParameter:
    name: videocardGpuVendor
    in: query
    type: string
    required: false

  VideocardVendorParameter:
    name: videocardVendor
    in: query
    type: string
    required: false